Move Inflector to Util
