+++ src/Annotation/AttributesHydratorTrait.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace ApiPlatform\Core\Annotation;
 
-use ApiPlatform\Core\Bridge\Doctrine\Inflector\Inflector;
 use ApiPlatform\Core\Exception\InvalidArgumentException;
+use ApiPlatform\Core\Util\Inflector;
 
 /**
  * Hydrates attributes from annotation's parameters.
