+++ src/Util/AnnotationFilterExtractorTrait.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace ApiPlatform\Core\Util;
 
 use ApiPlatform\Core\Annotation\ApiFilter;
-use ApiPlatform\Core\Bridge\Doctrine\Inflector\Inflector;
 use Doctrine\Common\Annotations\Reader;
 
 /**
